chore: storybook withNDsTheme addon was renedering stories too often, making animations invisible
